chore: remove outdated 404.html, favicon, index.html, and sitemap docs(mkdocs): update configuration for new theme & plugins
This commit removes several outdated files from the `docs/` directory, including the custom 404 page, favicon, index, and sitemap. Additionally, it updates the `mkdocs.yml` file to set up a new theme and add several extensions and plugins that improve documentation features.

The removal of these files might be part of an effort to clean up or restructure the existing documentation site. The update to the `mkdocs.yml` file suggests enhancements to the documentation process by enabling additional markdown features, such as tooltips, admonitions, and code annotations. This should result in a more interactive and user-friendly documentation experience.

Files removed:
- `docs/404.html`
- `docs/assets/images/favicon.png`
- `docs/index.html`
- `docs/sitemap.xml`
- `docs/sitemap.xml.gz`

Configuration updated in `mkdocs.yml`:
- Changed theme to `material`.
- Added new features like content tooltips.
- Enabled various Markdown extensions for extended syntax features.
- Updated navigation structure with new section titles and links.

These changes do not affect the core functionality of the project but rather its documentation and presentation aspects. Users and contributors should now experience improved navigation and readability when accessing the documentation.